Add files via upload
The objective of this web application is to make sure that farmers from all parts of the country are able to sell their products to clients from all regions within the country. It also going to make it easy for farmers to access both virtual and physical government services from all parts of the country. It is going to link farmers to agricultural experts and firms from where they can get equipment and other services.